re-factor orientation-related functions into new orientation module
